+++ CHANGELOG.md
@@ -40,6 +40,7 @@
     - updates css to use `flexbox` instead of the clever `translateY()` tricks.
     - Removes loading indicator and simplifies `LoadingView` [#885](https://github.com/openscope/openscope/issues/885)
 - Updates ENGM (Oslo Airport, Norway) [#796](https://github.com/openscope/openscope/issues/796)
+- Add contribution guidelines document [#723](https://github.com/openscope/openscope/issues/723)